engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "engine"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "asyncpg>=0.30",
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "sentry-sdk[fastapi]>=2.0",
]

# Required: project name is "engine" AND the package directory is "engine/".
# Without this, hatchling fails with "Unable to determine which files to ship".
[tool.hatch.build.targets.wheel]
packages = ["engine"]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.24"]
